Commit 24  -  25
Commit 24:

Added a custom "Not Found" error page to handle 404 errors. Introduced a new NotFoundIndicator component with styling and a message for missing pages. Updated the App.js to include routing for error handling.

Commit 25:

Implemented a "Welcome" page with a new WelcomeHeader component to display introductory content and an image. Updated routing in App.js to navigate to the welcome page. Improved the user interface with better header styling and page layouts.